Move LocalFeedImageDataLoader and FeedImageDataStore to production
